vmux-e1000: extract InterruptTHrottlerQemu -> interrupts/qemu.hpp
